requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-pytest==0.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
oauthlib==3.2.2
openai==0.26.5
packaging==23.0
Pillow==9.5.0
pluggy==1.0.0
psycopg==3.1.8
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.2
pytest-django==4.5.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
